Address PR #1052 review feedback
Automated feedback addressing for issue #1051.
