+++ packages/orchestrator/src/worker/repo-checkout.ts
@@ -7,6 +7,31 @@ import type { Logger } from './types.js';
 const execFileAsync = promisify(execFile);
 
 /**
+ * Default `git.remoteBranchExists` helper for the pre-push guard's production
+ * callers (#1051 FR-002; see `push-guard.ts`).
+ *
+ * Runs `git ls-remote --heads origin <branch>` in the given cwd (or the process
+ * cwd when omitted). Callers that need a different implementation (e.g. tests)
+ * inject their own via `PushGuardInput.git.remoteBranchExists`.
+ *
+ * Lives here rather than in `push-guard.ts` because `push-guard.ts` is
+ * deliberately kept free of `node:child_process` imports (repo-wide
+ * `no-restricted-imports` ban; #437). `repo-checkout.ts` is already on the
+ * allowlist and already owns every other `git` invocation in the worker.
+ */
+export async function defaultRemoteBranchExists(
+  branch: string,
+  cwd?: string,
+): Promise<boolean> {
+  const { stdout } = await execFileAsync(
+    'git',
+    ['ls-remote', '--heads', 'origin', branch],
+    cwd ? { cwd, encoding: 'utf-8' } : { encoding: 'utf-8' },
+  );
+  return stdout.trim() !== '';
+}
+
+/**
  * Manages git repository checkouts for workers.
  *
  * Supports two modes:
